### 2017-08-27 #381 @whedon
[REVIEW]: PlateReaderDataAnalyzer: a SAAS service to compute growth rates based on plate reader collected data
**Submitting author:** @uriba (<a href="http://orcid.org/0000-0002-3097-9681">Uri Barenholz</a>)
**Repository:** <a href="https://github.com/uriba/PlateReaderDataAnalyzer" target ="_blank">https://github.com/uriba/PlateReaderDataAnalyzer</a>
**Version:** v0.0.1
