### 2017-09-13 #381 @arfon
> @arfon Sorry for the delay. I meant to do this last week but ended up packing my house and evacuating because of Irma instead. I just got back home today - fortunately no major damage. I'll aim to complete the review next week.

Of course! Glad to hear you're safe (and the house isn't too bad). Please take your time and get to this when you can.
