### 2017-09-22 #381 @pboesu
Sorry again for the delay. Here's my first pass at this.

## General comments
This is a mostly functional shiny app that visualizes outputs from a proprietary plate reader. I do not work with plate readers, or associated software, so I can not evaluate the claim made in the rather brief paper.md, that this application is indeed easier to use than the cited alternative tools.
##Architecturee, documentation, testing
Generally, the application lacks documentation. There is some introductory text in the README, and the About tab of the app, but since the author does not make use of typical R package infrastructure *functions are undocumented and untested*.
I am aware that there is a distinction to be made between web applications and 'proper' R packages, but since the former can be packaged (see e.g. here https://will-landau.com/2016/11/01/appPackage/) the I would strongly prefer if @uriba would recast this as an R package, and use testing and documentation features, even if most functions are never to be invoked on the command line by the end user. This would also streamline the installation process.

Pop-up documentation is provided in the running app for some but not all drop-down lists, buttons, etc. at the very minimum these should be provided so the app user can understand the GUI.

## Functionality
The software appears incomplete, e.g. the "Plotly test" option in the "Data to display" drop down does not seem to trigger an action. Similarly, in the growth rate analysis setting the user can adjust maximum y-axis limits for some plots, but not others, and also there is no way to adjust minimum y-axis values. The growth rate analysis fails with "Error: incorrect number of dimensions" when the "Match" option is used with a single well. Such errors should be avoidable with adequate unit testing of the underlying shiny server functions.

## Software paper
I am aware that JOSS is not about verbose written descriptions. But I find two sentences rather too short. In particular the paper doesn't even mention the particular make and models of plate reader that this software is designed for.

## Other issues
mismatch in terminology: About panel talks about "plate layout file", interface has slot for "plate description file"

## Unticked list items
- Functionality: An error occurs when analysing single wells
- Performance: Unable to confirm, due to a lack of familiarity with the subject matter and alternative software solutions.
- Version: There is no github release
- Installation instructions: dependecies are stated in the readme, but the author sidesteps the R packaging process, which I would find preferable
- Functionality documentation: Is largely lacking, see general comments above.
- Automated tests: None.
- Community guidelines: None.
- References: reference `grofit` is missing a DOI

## Conclusions
Based on the above mentioned issues I recommend a *major revision*. I would really like to see this recast as an R package with appropriate documentation of both internal functions and GUI components. Ideally the author would also implement unit testing of the internal functions.
