### 2017-09-22 #381 @pboesu
@arfon Given that I feel like there's a need for a fundamental recasting of this app, is there a point in creating issues for all the points I raise?
