### 2017-09-25 #381 @arfon
> @arfon Given that I feel like there's a need for a fundamental recasting of this app, is there a point in creating issues in the target repository for all the points I raise?

Let's see what the author (@uriba) comes back with in response. I agree, it sounds like some fairly major enhancements/improvements are going to be necessary to this package before we can proceed further.
