### 2017-09-27 #381 @uriba
Dear @arfon and @pboesu,

@pboesu, many thanks for the thorough review!

I find @pboesu comments to be relevant and important to improve the quality
of the submitted software and make it more accessible for others to use
and/or contribute to.

I originally submitted this software for publication ~2.5 months ago in
order to get academic credit for work I've done during my PhD studies (but
that was not my main focus, nor of high, purely academic, importance, but
did serve, and continues to serve, fellow lab members as a useful utility),
a credit I needed for post-doc fellowship applications (which resonates
extremely well with the JOSS mission statement, as was detailed, in this blog
post <http://www.arfon.org/> by @arfon).

As the fellowships have all been submitted by now, my involvement in the
day-to-day lab work has decreased, and I currently face other obligations,
my capacity to fully address the issues raised in the review is limited.

If it is acceptable, I suggest allocating a 3 months period for me to
answer @pboesu 's comments and proceed from there. If I don't meet this
deadline, then the publication can be retracted.

I also suggest forming guidelines for the publication of SaaS services at
JOSS, as such services are different than 'classical' software packages in
terms of the technical capacity of their users (for example, not requiring
familiarity with code/installation etc.), testing methodologies,
versioning, etc. yet, given the recent advancements in cloud technology and
availability, I expect such software solutions to be increasingly adopted
as scientific open source software resources as well.

Many thanks again,

On Mon, Sep 25, 2017 at 9:53 PM, Arfon Smith <notifications@github.com>
wrote:

> @arfon <https://github.com/arfon> Given that I feel like there's a need
> for a fundamental recasting of this app, is there a point in creating
> issues in the target repository for all the points I raise?
>
> Let's see what the author (@uriba <https://github.com/uriba>) comes back
> with in response. I agree, it sounds like some fairly major
> enhancements/improvements are going to be necessary to this package before
> we can proceed further.
>
