### 2017-10-04 #381 @arfon
> If it is acceptable, I suggest allocating a 3 months period for me to answer @pboesu 's comments and proceed from there. If I don't meet this deadline, then the publication can be retracted.

@uriba - this sounds reasonable to me. Please let us know when you're ready to pick up this submission again.
