### 2018-06-16 #381 @uriba
Unfortunately, I don't see this happening anytime soon.
Moreover, given the advances in other web-oriented development platforms
(python+bokeh) I think better implementation options are currently
available.
This publication can therefore be dropped.
Thanks and have a nice weekend,
Uri Barenholz

On Fri, Jun 15, 2018 at 10:43 PM, Arfon Smith <notifications@github.com>
wrote:

> 👋@uriba <https://github.com/uriba> - any chance you might pick this up
> again sometime soon?
>
